requirements.txt
django
django-social-auth
django-annoying
django-debug-toolbar
